[Zodiac] Correct terminology (zodiac vs horoscope)
